Refactor

+++ src/Utils/FilesHelper.php
@@ -5,6 +5,7 @@ namespace Utils;
 use FilesystemIterator;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
+use SplFileInfo;
 
 class FilesHelper
 {
@@ -13,6 +14,8 @@ class FilesHelper
     public const string TEX_TEMPLATE = 'template/template.tex';
     public const string OUTPUT_BASE_PATH = 'output/';
     public const string TEMP_PATH = 'temp/';
+
+    private const string GIT_KEEP = '.gitkeep';
 
     public static function getDirectoryIterator(string $dir): RecursiveIteratorIterator
     {
@@ -30,7 +33,9 @@ class FilesHelper
         if (file_exists($dir)) {
             $ri = self::getDirectoryIterator($dir);
 
+            /** @var SplFileInfo $file */
             foreach ($ri as $file) {
+                if ($file->getFilename() === self::GIT_KEEP) continue;
                 $file->isDir() ? rmdir($file) : unlink($file);
             }
         }
